Move validators into their own directory
