+++ package/collection2/main.js
@@ -4,6 +4,7 @@ import SimpleSchema from "meteor/aldeed:simple-schema";
 import { EJSON } from 'meteor/ejson';
 import { flattenSelector, isInsertType, isUpdateType, isUpsertType, isObject, isEqual } from './lib';
 import { createSimpleSchemaAdapter, createZodAdapter, createAjvAdapter } from './adapters';
+import { getValidationContext, isZodSchema, enhanceZodSchema } from './validators';
 
 const meteorVersion = Meteor.release.split('@')[1].split('.');
 const noAsyncAllow = meteorVersion[0] >= 3 && meteorVersion[1] >= 1;
@@ -520,225 +521,8 @@ function doValidate({ collection, type, args = [], getAutoValues, userId, isFrom
         validationContext = schema.namedContext(validationContext);
       }
     } else {
-      // Special handling for Zod schema that might be missing namedContext
-      if (schema._def && schema._def.typeName === 'ZodObject') {
-        
-        // Create a validation context manually for Zod schemas
-        let errors = [];
-        
-        // Attach namedContext method if it doesn't exist
-        if (typeof schema.namedContext !== 'function') {
-          
-          // Define the namedContext method
-          schema.namedContext = function(name = 'default') {
-            return {
-              validate: (obj, options = {}) => {
-                try {
-                  if (options.modifier) {
-                    // For update operations with modifiers
-                    if (obj.$set) {
-                      // Create a partial schema for update validation
-                      const partialSchema = schema.partial();
-                      try {
-                        partialSchema.parse(obj.$set);
-                      } catch (error) {
-                        collectErrors(error);
-                        return false;
-                      }
-                    }
-                    
-                    // Handle other modifiers if needed
-                    if (obj.$push) {
-                      // We'll just allow push operations for now
-                    }
-                    
-                    if (obj.$unset) {
-                      // We'll just allow unset operations for now
-                    }
-                    
-                    // If no specific modifiers to validate, treat it as valid
-                    if (!obj.$set && !obj.$push && !obj.$unset) {}
-                    
-                    // For upserts, validate the combined document if provided
-                    if (options.upsert && options.extendedCustomContext && options.extendedCustomContext.upsertDocument) {
-                      try {
-                        schema.parse(options.extendedCustomContext.upsertDocument);
-                      } catch (error) {
-                        collectErrors(error);
-                        return false;
-                      }
-                    }
-                    
-                    // If we reach this point, all validation passed
-                    errors = []; // Clear errors if validation passes
-                    return true;
-                  } else {
-                    // For normal documents (insert)
-                    try {
-                      schema.parse(obj);
-                      errors = []; // Clear errors if validation passes
-                      return true;
-                    } catch (error) {
-                      collectErrors(error);
-                      return false;
-                    }
-                  }
-                } catch (error) {
-                  collectErrors(error);
-                  return false;
-                }
-              },
-              validationErrors: () => {
-                return errors;
-              },
-              invalidKeys: () => {
-                return errors;
-              },
-              resetValidation: () => {
-                errors = [];
-              },
-              isValid: () => {
-                return errors.length === 0;
-              },
-              keyIsInvalid: (key) => {
-                return errors.some(err => err.name === key);
-              },
-              keyErrorMessage: (key) => {
-                const error = errors.find(err => err.name === key);
-                return error ? error.message : '';
-              }
-            };
-            
-            // Helper function to collect errors
-            function collectErrors(error) {
-              errors = []; // Clear previous errors
-              
-              if (error.errors) {
-                error.errors.forEach(err => {
-                  errors.push({
-                    name: err.path.join('.'),
-                    type: err.code,
-                    value: err.received,
-                    message: err.message
-                  });
-                });
-              } else if (error.message) {
-                // Add a generic error if we don't have specific path errors
-                errors.push({
-                  name: 'general',
-                  type: 'error',
-                  value: null,
-                  message: error.message
-                });
-              }
-            }
-          };
-          
-          // Also add allowsKey method to the schema itself
-          schema.allowsKey = (key) => {
-            // For Zod schemas, check if the key exists in the shape
-            if (key === '_id') return true; // Always allow _id
-            
-            // Try to get the shape from the Zod schema
-            const shape = schema._def?.shape?.();
-            if (shape) {
-              return key in shape;
-            }
-            
-            // If we can't determine, default to allowing the key
-            return true;
-          };
-          
-          // Add a clean method for Zod schemas
-          schema.clean = (obj, options = {}) => {
-            const { mutate = false, isModifier = false } = options;
-            
-            // If not mutating, clone the object first
-            let cleanObj = mutate ? obj : JSON.parse(JSON.stringify(obj));
-            
-            if (isModifier) {
-              // For update operations with modifiers like $set, $unset, etc.
-              if (cleanObj.$set) {
-                // Process each field in $set
-                Object.keys(cleanObj.$set).forEach(key => {
-                  const value = cleanObj.$set[key];
-                  
-                  // Remove empty strings if option is enabled
-                  if (options.removeEmptyStrings && value === '') {
-                    delete cleanObj.$set[key];
-                  }
-                  
-                  // Auto-convert strings to numbers/booleans/etc if option is enabled
-                  if (options.autoConvert) {
-                    // Would need more complex logic to properly convert types
-                    // For now, we just do a basic conversion for common types
-                    if (typeof value === 'string') {
-                      if (value === 'true') cleanObj.$set[key] = true;
-                      else if (value === 'false') cleanObj.$set[key] = false;
-                      else if (!isNaN(value) && value.trim() !== '') cleanObj.$set[key] = Number(value);
-                    }
-                  }
-                  
-                  // Trim strings if option is enabled
-                  if (options.trimStrings && typeof value === 'string') {
-                    cleanObj.$set[key] = value.trim();
-                  }
-                });
-                
-                // Remove $set if it's empty after processing
-                if (Object.keys(cleanObj.$set).length === 0) {
-                  delete cleanObj.$set;
-                }
-              }
-              
-              // Process other modifiers if needed
-            } else {
-              // For insert/update operations without modifiers
-              
-              // Process each field in the document
-              Object.keys(cleanObj).forEach(key => {
-                const value = cleanObj[key];
-                
-                // Remove empty strings if option is enabled
-                if (options.removeEmptyStrings && value === '') {
-                  delete cleanObj[key];
-                }
-                
-                // Auto-convert strings to numbers/booleans/etc if option is enabled
-                if (options.autoConvert) {
-                  // Would need more complex logic to properly convert types
-                  // For now, we just do a basic conversion for common types
-                  if (typeof value === 'string') {
-                    if (value === 'true') cleanObj[key] = true;
-                    else if (value === 'false') cleanObj[key] = false;
-                    else if (!isNaN(value) && value.trim() !== '') cleanObj[key] = Number(value);
-                  }
-                }
-                
-                // Trim strings if option is enabled
-                if (options.trimStrings && typeof value === 'string') {
-                  cleanObj[key] = value.trim();
-                }
-              });
-            }
-            
-            return cleanObj;
-          };
-          
-          // Set default clean options
-          schema._cleanOptions = {
-            filter: true,
-            autoConvert: true,
-            removeEmptyStrings: true,
-            trimStrings: true
-          };
-        }
-        
-        // Set the validation context to use the schema's namedContext
-        validationContext = schema.namedContext();
-      } else {
-        validationContext = schema.namedContext();
-      }
+      // Use the getValidationContext helper to handle different schema types
+      validationContext = getValidationContext(schema);
     }
 
     // Add a default callback function if we're on the client and no callback was given
